Avoid some -Wconversion warnings by changing parameter types
These functions can safely handle bigger types, so
avoid needless downcasts.

+++ openvpn/common/string.hpp
@@ -413,7 +413,7 @@ inline std::string reduce_spaces(const std::string &str, const char rep)
 }
 
 // generate a string with n instances of char c
-inline std::string repeat(const char c, int n)
+inline std::string repeat(const char c, size_t n)
 {
     std::string ret;
     ret.reserve(n);
@@ -423,7 +423,7 @@ inline std::string repeat(const char c, int n)
 }
 
 // generate a string with spaces
-inline std::string spaces(int n)
+inline std::string spaces(size_t n)
 {
     return repeat(' ', n);
 }
